api/requirements.txt
alembic==1.9.2
Brotli==1.0.9
cffi==1.15.1
click==8.1.3
cssselect2==0.7.0
Flask==2.2.2
Flask-Cors==3.0.10
Flask-Migrate==4.0.3
Flask-SQLAlchemy==3.0.3
flask_weasyprint==1.0.0
fonttools==4.38.0
greenlet==2.0.2
html5lib==1.1
importlib-metadata==6.0.0
itsdangerous==2.1.2
Jinja2==3.1.2
Mako==1.2.4
MarkupSafe==2.1.2
numpy==1.24.1
pandas==1.5.3
Pillow==9.4.0
psycopg2-binary==2.9.5
pycparser==2.21
pydyf==0.5.0
pyphen==0.13.2
python-dateutil==2.8.2
python-dotenv==0.21.1
pytz==2022.7.1
six==1.16.0
SQLAlchemy==2.0.1
tinycss2==1.2.1
typing_extensions==4.4.0
webencodings==0.5.1
Werkzeug==2.2.2
zipp==3.12.0
zopfli==0.2.2